adressing CI comments
