+++ Core/src/Detector/CMakeLists.txt
@@ -2,6 +2,7 @@ target_sources(
   ActsCore
   PRIVATE
     detail/BlueprintHelper.cpp
+    detail/BlueprintDrawer.cpp
     detail/CylindricalDetectorHelper.cpp
     detail/PortalHelper.cpp
     detail/SupportHelper.cpp
